Restructured test_fread_large for cleaner output when testing.
All tests in this file will be skipped if environment variable `DT_LARGE_TESTS_ROOT` is undefined or an empty string.

A hack involving fixtures was used to give each test the ability to run and report as individual instances. Undefined paths now report as an instance error and does not abort the entire test process. (#248)