Added ability to test `test_fread_large.py` when `DT_LARGE_TESTS_ROOT` is defined (#256)
* Added 'Determine Environment Variables' stage which assigns a value to environment variable DT_LARGE_TESTS_ROOT only if fread-type files were modified.

Created environment variable DT_LARGE_TESTS_ROOT in osx/linux test stages.

Added 'module' argument to `make test` calls to implement additions made to Makefile. (#248)

* Added `test_install` command that installs all modules specified in `extra_requires[testing]`.

Added `MODULE` var argument for use in `test_install` which specifies the datatable module to be used (default is `.`).

Added `make test_install` as a step in `make test` and `make coverage`. (#248)

* Changed minimum version requirement for pytest to 3.1 (#248)

* Restructured test_fread_large for cleaner output when testing.

All tests in this file will be skipped if environment variable `DT_LARGE_TESTS_ROOT` is undefined or an empty string.

A hack involving fixtures was used to give each test the ability to run and report as individual instances. Undefined paths now report as an instance error and does not abort the entire test process. (#248)

* Moved stage "Determine Envrionment Variables" into a function (#248)

* Skipped files in directories will be marked as skipped in pytest instead of just printing as message