WIP
Created Stats structure to store column statistics in a DataTable. The DataTable structure now contains an array of Stats structures. Each Stats structure contains the statistics of the column at the same index.

All modifications to a Stats structure should be done through the functions provided in `stats.h`. The following functions were implemented:
    - Initialize, free, and copy a Stats structure.
    - Determine whether a statistic has been computed for a Stats structure.
    - Compute a statistic for every Stats in a DataTable, if possible.
    - Create a new DataTable that contains the value of a statistic for every
      column in a DataTable.

Currently the only statistics available are min and max for boolean, integer, and float types.

TO DO:
    - Integrate Stats into DataTable operations
    - Provide support for RowIndexes
    - Check for Stats integrity in datatable_check.c
    - Expose to Python
    - Unit tests
    - Use OpenMP for computing statistics
(#275)