Implement .to_csv() function (#364)
This adds functionality to write a DataTable into a csv file:

    import datatable
    dt0 = datatable.DataTable(...)
    dt0.to_csv(filename)

If filename is omitted, then the csv text will be returned as a string.

This is a bare-bones functionality, no additional options are supported. Nevertheless, it is supposed to be very fast. In particular, the writing is done in multi-threaded environment; all writers are individually optimized (see micro benchmarks); double values are written in hex format (later this will become an option, which will be turned off by default). Strings are auto-quoted, and any unsafe symbols are escaped.

There is a tiny chance for the process to seg.fault while writing, but this is unlikely to happen for any real-life dataset. Later we will add more robust size checks, to eliminate even this tiny possibility.