Refactored `verify_integrity` to be modular. (#479)
* Refactored `verify_integrity` to be modular. Changed the parameters to use a `vector<char>*`
instead of `char**`, and added a `max_errors` and `name` parameter. `max_errors` sets the
limit for how many error messages should be printed. `name` is the string to use when
referring to the class instance in error messages.

The `verify_integrity` method is a public method in every class. Column's also have a
(protected) `verify_meta_integrity` which does integrity checks to the meta array.

Some additional commenting may be needed. Additionally integrity checks for `Stats` need
to be implemented (and for MemoryBuffer?)

(#333)

* Typo/bug fixes. Applied suggestions made in comments
(#339)