Memory Map Manager (#570)
Introduce the MemoryMapManager class, which holds information about all (or most) memory maps currently held by the process. If we ever run into "errno 12" (not enough memory) when opening a memory map, this class has the ability to request some of the entries to become unmapped, allowing new entries to be created. This should eliminate most of the "Errno 12" errors.