`StringColumn`s now use two `MemoryBuffer`s: one for its offsets (`mbuf`) and one for its strings (`strdata`). The use of an offset field or padding is no longer needed. (#583)
Every offset buffer (retrieved through the `offsets` method) begins with a `-1` value. Thus it should be impossible for a `StringColumn` to have an empty offset buffer. This also implies that the buffer will have a length of `nrows` + 1.
The initial `-1` value is ignored as a logical entity in the `StringColumn`, meaning that `data_nrows()` will return `nrows` instead of `nrows` + 1. The `offsets` method also returns a pointer to the value after the `-1` (making `offsets()[-1]`) a valid statement.

The string buffer is retrieved through the `strdata` method, but the resuling pointer points to the value BEFORE the intial character. Thus, `strdata()[0]` may result in a segfault.

Other notable modifications:
  - String columns are stored over two files: file `X` contains the offset data and file `X_str` contains the string data. The `_str` extension is necessary and should never be removed.
  - The `StrBuf` struct used in `py_fread.c` now uses `MemoryBuffer` references instead of a character array. This allows for the algorithm to write strings directly into the destination column instead of storing it inside a temporary file.
  - A result of the above change is that some `goto` labels are never reached. I wrapped the affected functions in try-catch blocks (The macros in `exceptions.h` cannot be used due to the need of some post-error cleanup).
  - Sorting methods have gotten uglier due to the need to cast columns into a `StringColumn`. Additional "if" statements were written to account for the new structural differences in the `StringColumn`s. We should look into modularizing the `sort` method.

(#552)