Introduce "GenericReader" class in C (#631)
The primary purpose of the new class is to provide a common starting ground for various file readers. The fread reader is one example, arff reader or fw reader are other (coming later). The new class also unifies management of lifetime of different variables used during the parsing.

This PR also contains beginning of the code for Arff Reader implementation. This is currently disabled, merging in only to avoid horrible merge conflicts later on.