Chunked reader (#667)
Multiple updates to ChunkedReader. It now supports limiting the number of rows read, and deal with the case when the number of rows preallocated is lower than necessary.

This functionality is currently orphaned.

Also the default value of parameter skip_blank_lines was changed to False: this can better address a "footer" at the end of a csv file.