Split logic in fread into separate functions (#757)
This is a preparation step towards using ChunkedDataReader in fread, and towards solving some of the pressing issues regarding bad chunking behavior.
Note that the current result is an uglier version of fread, with worse diagnostic information. This PR should be considered an evolutionary step towards better, brighter future. Nevertheless I believe this should be merged so that we can see better how the code changes in the future, and in order to reduce the size of the final change. Plus, this will allow us to solve some of the outstanding issues...