Categorical aggregations, except for the 2D case.
Initial implementations of
	- 1D categorical aggregation
	- 2D mixed aggregation

From now on, the original dataframe will not be modified in-place,
instead, we return a new dataframe consisting of the shallow copies of
all the aggregated columns and include the binning information as the
additional one at the end. Only int32_t bins are supported, because of
the same restriction being valid for `groupby`.