String representation (#1132)
Change internal representation of NA strings: now string offsets are assumed to be uint32_t / uint64_t, and NA strings are marked with the highest bit set in the end offset of the string. The first element in the offsets array is now always 0, and the offsets are assumed to be from the beginning of the string buffer (not from the -1 byte as before).

This change affects how the string data is stored internally, and therefore also how it is saved in the NFF files. However, the change is backward-compatible in the sense that if an old NFF file is opened with dt.open(), the old format of strings will be automatically converted into the new one. This may cause a small amount of slowdown, as the data for all string columns needs to be loaded into memory and then processed. When saving, such data will be saved in the new format (there is no option to save in the old format).