Python framework, continued (#1230)
This PR introduces further improvements towards implementing the new Python-C++ framework. Although it is far from complete, I anticipate that further changes will be done on case-by-case need, and there won't be a point in time when we could definitely say "the framework is ready, now the issue can be closed".