Implement fuzzy column name matching (#1259)
Whenever a column is selected by name from a Frame, and such name doesn't exist in this Frame, we will now report column names that are sufficiently similar. This allows the user to fix common spelling mistakes more easily.
The runtime overhead is only paid when the error actually occurs, and thus it should be considered negligible.

The algorithm used to find approximate matches is based on Levenshtein distance. The algorithm is slightly modified to allow the cost of substitution to depend upon the character classes. In particular, we treat case-mismatch as a much smaller error than changing one letter into another, which is again smaller than changing a letter into non-letter.