Construct Frame from a list of tuples (#1264)
This adds the ability to create Frame from rows of data, where each row is represented as a `tuple` or as `namedtuple`. Thus,
```
    dt.Frame([(1, "foo", 3.7)])
```
creates a 1x3 Frame with columns containing `[1]`, `["foo"]` and `[3.7]`. At the same time,
```
    dt.Frame([[1, 2, 5]])
```
still creates a 3x1 Frame with 1 column containing `[1, 2, 5]`.

The distinction comes from the fact that in many strictly typed programming languages lists usually have homogenous types (and therefore are best suited to represent a column of data), whereas tuples can be collections of heterogenous types, which is what needed to represent a row in the data.

Closes #1263