Randomized integration testing framework (#1286)
This adds 2 new scripts:
* `tests/random_driver.py`
* `tests/random_attack.py`

The first one is the main one. Run it manually in order to perform a "random attack"-style integration testing. The script will invoke "random_attack.py" sequentially, with different initial seeds. Failed runs will be reported. The system is sufficiently robust to handle segfaults/freezes in the attack script.

Each execution of random attack starts with a random seed; generates a frame of random data; then performs a sequence of randomly chosen transformations; and finally checks whether the final result is correct. The correctness is verified by keeping side-by-side a datatable Frame, and a pure-python "frame", and performing all transformations on both frames simultaneously.

This PR establishes the framework and implements 5 transformations. More methods will be created in subsequent PRs.

Closes #1277