Split string and nhot-encode (#1304)
This PR adds a new function dt.split_into_nhot(frame, sep), which acts upon an n x 1 Frame (of str32 or str64 type) and produces an n x k Frame with boolean columns. Each value in the input column is split according to the provided separator, the whitespace is trimmed, and the resulting pieces are converted into individual columns in the output frame.