Improve appearance of example code / results in documentation (#1342)
The documentation must include both code snippets and their results. This module mostly produces data frames as the result, which is a styled HTML table. Hence, the results must be included in the .rst files in the form of raw HTML, with an accompanying stylesheet to give the table proper appearance.

This PR:
* Adds `conf.py` and a makefile for configuring the generation of the documentation;
* Adds `_static/code.css` file to style the datatable outputs in the documentation, and to give it more Jupyter-like appearance; 
* tweaks the `_repr_html_()` method so that its result is easier to include into an .rst file;
* modifies `using-datatable.rst` to provide output examples for the first snippet(s) of code.