Do random sampling for categorical data sets in aggregator (#1346)
For 1D and 2D categorical aggregators we may end up with too many exemplars being returned because binning for categoricals is essentially based on `datatable` grouping. Also, for ND aggregator, when there are no continuous columns, all rows are becoming exemplars, that may be too much for `vis-data-server` if a data set is large. 

One of the options to reduce the number of returned exemplars is to do sampling and return `n_bins` random exemplars for 1D, `nx_bins * ny_bins` random exemplars for 2D and `nd_max_bins` random exemplars for ND aggregator.

Closes #1345