Allow datatable to be compiled with gcc on linux (#1362)
This PR tweaks certain compiler/linker flags, as well as fixes few compilation errors (that inexplicably are perfectly fine with Clang, not even a warning raised). 

The biggest change to code structure is that now python module-level functions are declared using a macro rather than a template. Even though this strays away from C++ ideals, the macro-based solution turns out to be cleaner and the code smaller. In addition, the function's signature is now defined together with the function itself, which makes the code more coherent.

The `setup.py` now does the work for setting up a compiler only when it detects a command that will cause an extension to be built.

With this PR I was able to build `datatable` on Ubuntu 18.04 without downloading Clang/Llvm. The build steps were the following:
1. set up python virtual environment
2. check out this branch from git
3. `$ make build`

Closes #1100