Implement FTRL-Proximal online learning algorithm (#1390)
* Initial implementation of FTRL method

* Adding OpenMP and feature interaction support

OpenMP is lock-free, i.e. all threads can in principle update the same
weight at the same time. However, probability of this is very small,
because number of features should be much less than the number of bins.

Feature interaction still to be thoroughly tested. Doesn't seem to work
well on datasets with large number of columns. For small number of
columns works pretty well.