Deprecate DT[col] (#1407)
As outlined in #1395, syntax DT[col] will now be deprecated. Please use DT[:, col] instead.

Starting from 0.9.0, DT[x] will be interpreted as row-selector instead. For example, DT[:5] will be the first 5 rows of the Frame, and DT["A"] will return the row keyed by the value "A".

This also fixes bug #1406 where empty tuple used as a selector caused a crash.