Add methods to_tuples() and to_dict() (#1439)
This PR adds 2 new Frame methods `.to_tuples()` and `.to_dict()`, and replaces the `.to_list()` method with the one implemented in C++.

`Frame.to_tuples()` converts the frame into a list of tuples, where each tuple represents one row. Note that we already have the ability to convert a list of tuples into a Frame, this is just the inverse of that operation: `list_of_tuples == dt.Frame(list_of_tuples).to_tuples()`.

`Frame.to_dict()` is similar to `.to_list()`, except that we return a dictionary of columns (keys are column names) instead of a list. This operation is also an inverse of creating a Frame from a dictionary: `dict_of_lists == dt.Frame(dict_of_lists).to_dict()`.

Closes #1400