Traverse exemplars in a quasi-random order in ND aggregator (#1553)
When exemplars are traversed in the order they appear in the `exemplars` vector, first exemplars will always have a priority over the last ones, hence, first exemplars "absorb" most of the members. In this case one can not distinguish real outliers with the exemplars (the last ones) that just had no chances to get some members.

One of the ways to fix it is to traverse exemplars in a random order, so that each exemplar gets the "same" chances to "absorb" a particular member. Here we use modular arithmetic to traverse exemplars in quasi-random order, as this approach seems to have negligible overhead, when comparing to `std::random_shuffle`, and we do not really need to be fully random, but just different enough to not inflate some particular exemplars.