max/min for an empty column now returns None (#1625)
This PR changes the behavior of functions `max()` and `min()` when they are applied to empty or all-NA columns. Previously those functions returned `-Inf` and `+Inf` respectively (or, for integer columns, `-MAX` and `+MAX`). Now these functions will return `None`.

The reasons for the change are discussed in #1624, but in summary they are:
- in mathematics max and min of an empty set do not exist, only supremum and infimum;
- for integers returning a large integer looks like a bug, and we prefer the behavior of floats to match the behavior of integers;
- numpy returns nan for max/min of array of nans (with a warning);
- R returns NA for max/min of a vector of NAs (however, it returns ±∞ if option `na.rm=TRUE` is used).

Closes #1624