Getters / setters revamp (#1658)
This PR changes changes the way getters / setters are declared, similar to #1657.

Now in order to declare a getter [+setter], you have to create a static `GSArgs` variable (similar to `PKArgs` for methods). This variable only carries the name of the property and its docstring. The advantage of the new approach is that now the documentation for a property can be declared right next to the property itself, which is a big plus.
In addition, the amount of generated code is now smaller, and we don't use templates as heavily. Instead, there are now 2 macros:
```
ADD_GETTER(gs, &Class::getterfn, args);
ADD_GETSET(gs, &Class::getterfn, &Class::setterfn, args);
```