Properly display keyed frame in console view (#1676)
This PR fixes a bug where the key columns were not displayed quite right during scrolling in a console widget. Now they have proper appearance of "frozen" columns (similar to Excel).

In addition, `Frame._data_viewer()` method is removed; instead there is now a helper class `DatatableFrameConnector` that has the same purpose. The advantage of having it as a separate class is that we could in principle add more classes like this to implement connectors to other data frames as well.

Closes #1672