Prevent overflow when constructing str32 columns (#1696)
Direct construction of StringColumn<T> objects is now prohibited (except for the cases when a blank column is created). Instead, we use new_string_column() function which returns either StringColumn<uint32_t>* or StingColumn<uint64_t>*, whichever is more appropriate.

This function also takes care of "fixing" the offsets array in case it overflown during the construction. This works provided that no individual string in the column exceeds 2Gb in size.

Closes #1694