Upgrade the wheel URL in requirements_docs.txt (#1719)
Until #952 is implemented, we cannot have a proper "latest" version of documentation on RTFD.io. What we can do, for now, is to manually update every once in a while the URL of the latest wheel in "requirements_docs.txt".

Also it appears that RTD site has switched to using Python3.7, so we need to update the wheel for that reason as well.