Fix creation of Frame in fread for str64 columns with NAs (#1841)
So the problem was that the code in fread that handled conversion of str32 -> str64 columns was not taking into account NA values properly. This was an old code (which means the issue existed for a very long time).
Today we have a similar code in StringColumn class itself, that handles the conversion correctly. Thus, simply removing the fread's code solved the problem: the conversion now happens automatically in the StringColumn class.

This PR also patches up a minor MacOS bug with fcntl() function sometimes returning EINVAL error even if it shouldn't. This bug was introduced yesterday, and it's unclear how it passed the unit tests. The "solution" is just to ignore this case, since the bug was fixed in the most recent version of MacOS already.