Create `.. dtframe` directive for documentation (#1861)
This directive will allow us to render frames in the documentation from easily-readable csv-like form in an rst file.

See documentation on how to use this new directive in file "dtframe_directive.py", and examples in the .rst files updated with this PR.

The module `datatable.sphinxext.dtframe_directive` depends on `docutils`, however this does not become the dependency of `datatable` itself, since `datatable` does not import that module. The dependency arises only when someone tries to import `datatable.sphinxext.dtframe_directive` directly to use with Sphinx -- however, in that case they would already have `docutils` on their systems, since it is a prerequisite for `sphinx`.