Virtual columns (#1865)
This PR adds new functionality: lazy computation of expressions via virtual columns.

Consider a columnar expressions such as `f.A * exp(f.B / 2)`. An eager evaluator computes this expression roughly as follows: first, column `B` is extracted from the frame, then it is cast to `float64` type, then the `/` operator divides by 2.0 and stores into a temporary column, then we compute `exp()` of that column (creating a new column), then we extract column `A` and cast it into `float64` too, and finally compute the product of the latter 2 columns. As you can see, this whole process involves creating many intermediate columns, which affects performance and memory consumption.

The new "lazy" evaluator will instead do the following: first a virtual column for `B` is created, then we create a virtual column for `float64(B)`, then a virtual column for `float64(B) / 2`, and so on, until finally we have a virtual column for the entire expression. Only then we start "materializing" the virtual column: compute the entire expression for each row, and store the result into memory, just once.

Thus, a virtual column is just a class which knows how to compute the value at a specific index `i`.

This "virtual columns" framework has a number of potential advantages over the eager evaluation:
- No intermediate memory used during the calculations;
- Possibly better performance, since most operations are memory-bound not CPU-bound, so avoiding extra memory writes should in theory allow the final result be computed faster;
- The code can be extended more easily compared to the traditional "eager" mappers;
- String columns are easier to work with too;
- The framework can be extended to allow virtual columns to be stored in a Frame directly (without materializing);
- The virtual columns can also be used to implement compressed data storage (see #1748).

This PR includes the following functionality:
- `virtual_column` base class;
- virtual <-> real column conversion;
- cast functions for virtual columns;
- unary operators for virtual columns;
- literals (basic implementation only);

This does **not** include yet:
- binary operators for virtual columns (just started working on an implementation);
- ability to include virtual columns in a frame;
- switch between lazy / eager evaluation mode (thus, the whole framework is "not operational");
- tests.