Fix segfault in progress reporting (#1879)
In some cases `progress_manager` tries to access progress bar that is already deleted by `progress_manager::finish_work()` method, and this results in a segfault. In particular, some of the aggregator tests, that also tested progress reporting, segfaulted on Linux and PPC.

To fix it, we use `std::mutex` to protect progress bar access.