Do binomial check in label_encode functions only when adding new labels (#1892)
For binomial problem number of labels should be less or equal to two. This check should only be performed when a new label is encountered in the target column. When running label encoding in parallel, we now do the check making sure the label was not added from another thread.

Closes #1891