Iterable frame (#1908)
This PR makes `Frame` object iterable:
```
for column in frame:
    # do something with `column`
```
The iteration in reverse order is also supported:
```
for column in reversed(frame):
    # do something with `column`
```

Note that unlike `pandas`, in our case iteration yields the entire column, not just the name. Obviously, the column is just a Frame with ncols=1, so its name can be retrieved trivially: `column.names[0]`.

This PR also addresses a bug with `cbind()` (which didn't handle temporaries quite well). The following invariant now holds:
```
DT == dt.cbind(*DT)
```

Closes #1907 
Closes #1906 
Closes #1905