Upgrade CSV writer (#1924)
This PR re-architects the internals of the CSV writer. The changes are aimed at increased extendability, and with an eye on virtual columns that will be added in 0.10. Check the description in write_manager.h for more information on the new architecture.

The issue that caused crashes excessive memory usage (#1921) was also resolved.

Closes #1921