Beginning to refactor the Column class (#1934)
The need for this refactoring is explained in issue #1396. 
This PR is incomplete (i.e. it doesn't covers the scope of #1396 fully), yet it makes an important first step towards achieving this goal. 
I'm proposing this as a PR, instead of continuing working on the issue, because it already grew in size to almost 1000 lines, and it could become unbearable to review if we wait until the issue is implemented completely.

The main idea of this PR is that a new class `OColumn` is added (in the future will be renamed into simple `Column`), meaning a column that is "owned". The `DataTable` class uses these objects as the primary storage of columns. Currently the code is able to compile & test successfully.

WIP for #1396