Cast to floats on-the-fly in Aggregator (#1938)
As we had a performance issue calculating stats on integer columns, some time ago Aggregator was switched to cast all data to float columns and calculate stats on the casted columns only. Though this approach made Aggregator faster, it obviously introduced a significant memory overhead (~x2).

As of #1894 most of the performance issues connected to stats are solved, so now we can do casting on-the-fly in Aggregator. With this PR we decrease memory usage by about factor of two, as well as slightly improve overall Aggregator performance.

Thanks to @st-pasha, in this PR we fix minor glitch with progress bar rendering too. Testing progress bar capabilities further revealed problem #1937, that should also be fixed at some point.