Column API cleanup (#2039)
`Column::get_valid()` methods now return `true` if the returned value is valid, and `false` if it is missing (NA). Before we returned the opposite: true if missing, false if valid. This change brings us closer to having better interoperability with Arrow.

Declared (but not fully implemented) tentative API for buffer access in `Column` class.