Convert python integers into int32/int64 only during frame creation (#2127)
Before this PR, `dt.Frame([1, 2, 3])` would create an `int8` column, and `dt.Frame([100, 200, 300])` had stype `int16`. After this PR both examples will create an `int32` column instead.

The reason for this change is that it is too error-prone: the user may not notice that int8/int16 columns are created, and cause an integer overflow during simple mathematical operations. With int32 the likelihood of an overflow is much less.