Improve stylesheet injection into a Jupyter notebook (#2132)
- Added method `datatable.init_styles()` which injects datatable's stylesheets into a Jupyter notebook;
- This method is also run automatically when datatable is imported.

Note that this PR does not solve the problem of disappearing stylesheet entirely: if have a cell `import datatable as dt`, then the stylesheet will be automatically added when the cell is executed, but then will disappear if the user runs the same cell again. In a situation like this the user would have to manually call `dt.init_styles()` to restore the stylesheet. This is far from ideal, because a normal user would probably not even know about the existence of this function...

I've experimented with different approaches to this problem, but neither worked:

- Overriding Python import hooks, forcing python to re-import datatable each time (or at least, run a custom function each time). This seems like it could have worked, but unfortunately there are too many places where datatable could get imported (including within its own code), and we may end up producing copious amounts of CSS code;

- Creating an output node with custom `__del__` method, so that it would revive whenever the node gets deleted. Unfortunately, IPython is not storing these nodes anywhere after they get displayed, which means it doesn't know what nodes still exist in Jupyter and which nodes got deleted.

- Creating a Javascript snippet which sets up an event handler when the CSS node gets removed, and then automatically restores that node. This would work within a notebook webpage (provided the user hasn't disabled javascript in notebooks), but breaks as soon as the notebook is stored as an ipynb file: Jupyter will not save any DOM elements that we "sneakily" added.

In view of these difficulties, I opted to use the explicit function method, which will at least work reliably for users who care about such display issues.

Closes #1871