Terminal string enhancement (#2139)
This PR enables printing of properly formatted nested strings into the Terminal.

So, a primary problem with using ANSI escape sequences is that they do not nest: for example if you want to print `bold("TEST " + grey("...") + " STRING")`, then the `grey()` function will emit a RESET code at the end, which will both revert to the main color (as intended), and remove the "bold" attribute (not intended).

Thus, in order to properly render a formatted string into a terminal, the string cannot be "cooked": it must remember which parts carry which style attributes, and only apply those when it is rendered into the final output stream. This is achieved now via the `tstring` class, together with the `TerminalStream` class.

The `Terminal` class no longer has any text-styling/coloring capabilities, the same functionality should now be achieved via the `tstring` class, and then sending it into a `TerminalStream`.