Added parameters append= and header= to .to_csv() (#2143)
Method `Frame.to_csv()` gains two new boolean parameters: `header=` and `append=`. The `header=` parameter controls whether or not to write into the output the header row with column names. The `append=` parameter allows the CSV content to be appended to an existing file instead of overwriting it:
```python
DT.to_csv("out.log", append=True)
```

By default, `append=False`. The default value of `header=` parameter is inferred: it is equal to `True` in all cases except when both `append=True`, and the file given as the argument already exists.

Closes #2130