Function .len() now returns the number of codepoints in a string (#2187)
Note that this does have an effect on performance. Unfortunately, there is no other way to make this function return correct result.

In the future we may keep a stat indicating whether the column is ASCII or not. And in case it is ascii, use the old simpler function for computing string length.

Closes #2160