Fix segfault when C++ tests were called with missing arguments (#2197)
Even though we do not expect users to call internal `datatable.lib.core` C++ tests, they should not segfault when called with missing arguments. In this PR we add a `PKArgs::check_required_args()` method that checks if the required arguments are missing and improve`test_parallel.py` to make sure C++ tests do not segfault.