Move monitor thread from idle_job to thread_pool (#2203)
Creating monitor thread on the datatable import causes a deadlock on Windows, see this [discussion](https://stackoverflow.com/questions/32252143/stdthread-cause-deadlock-in-dllmain) for more details. In fact, we don't really need the monitor thread being available right after the `import datatable`, so its creation can be easily postponed. 

In this PR we 
- move the monitor handler from `idle_job` to `thread_pool` class, and create the monitor thread only when it is needed by datatable;
- disable fork related code on Windows, because there is no fork mechanism there.

WIP for #1369