Do not use stdin and stdout as variable or function names (#2207)
In MSVC `stdin` and `stdout` are not valid variable and function identifiers. Even doing something like `void* stdinptr = ::stdin;` from a namespace leads to `C2589` compilation error. 

In this PR we rename all the datatable identifiers with the name of `stdin` to `rstdin` and `stdout` to `rstdout`, because they're representing `robj`s. We also replace `::stdin` with just `stdin` to make it compatible with MSVC. In future we should avoid using these identifiers in datatable, in such a case `::stdin` and `stdin`, `::stdout` and `stdout` will be equivalent.

WIP for #1369