Replace custom wheel URLs with pip install in README (#2228)
Right now the links to install datatable on Linux, that are included in `README.md`, are broken for all the python versions. This brings a lot of confusion, because this information is also shown on the main github page for datatable. In this PR we replace installation from these links with a single `pip install datatable` command, that will install datatable v0.10 on Linux and Mac. When custom links are fixed, we can include them under the detailed installation instructions in the docs.