Fix test_fread_omnibus (#2233)
`tests/fread/test_fread_random.py::test_fread_omnibus` is failing for some seeds due to the fact that `generate_str_column()` generates a column filled with only numeric values, for example for the seed of `1795109353` the column data is `['7E1', '']`. When `dt.fread()` receives such a data it converts it to a float column, while in the test the type was assumed to be `ltype.str`. 

In this PR we add a check to `generate_str_column()` function to make sure it never generates fully numeric data.