Added Sphinx directive ..ref-context (#2288)
This new directive makes it easy to link to functions / methods from datatable doc pages. In particular, the following functionality is enabled:
```
.. ref-context:: datatable.Frame

* :meth:`.to_dict()`
* :meth:`Frame.to_tuples`
* :meth:`.to_csv(filename, **kwargs)`
```
these will all be linked to their proper targets: `datatable.Frame.to_dict`, `datatable.Frame.to_tuples` and `datatable.Frame.to_csv` respectively. At the same time, the links will retain their text as written, allowing us to tailor the text to the context where it appears. 

For example:
```
*  Added function :func:`repeat(frame, n)` that creates a new
   Frame by row-binding ``n`` copies of the ``frame``.
```
Here it makes sense to write the function with its list of parameters, so that we could also describe those parameters in the same sentence.