Add dt.options.progress.allow_interruption option (#2300)
Add the `dt.options.progress.allow_interruption` option that will control if datatable is allowed to set a `SIGINT` handler or not. This option is `True` by default. If this option is set to `False`, then it won't be possible to interrupt any long-running datatable tasks with `Control+C`.

WIP for https://github.com/h2oai/h2oai/issues/13469