Big Move (#2305)
In this PR we relocate all source files -- both C++ and Python -- into the `/src` directory.
The reason for why this setup is preferred is explained in https://blog.ionelmc.ro/2014/05/25/python-packaging; however for us such move is in fact a necessity: with our previous setup it was impossible for `pytest` to properly test datatable installed from a wheel. Although pytest may eventually fix this problem (https://github.com/pytest-dev/pytest/issues/4714), for now the most reliable solution is to have the `datatable/` folder not in the root of the repository.

As for the `c` folder, there was no good reason to move it, except that now that we have the `src/` folder it seems only logical that the C++ code should also live there.

Also, I've inserted `sys.path.insert(0, "src")` in the tests/__init__.py and tests/conftest.py -- these are temporary changes, designed to ease the transition towards the new file layout.